[fact] Move the GumTreeSpoonAstDiff class to the util package #29
