+++ src/main/java/se/kth/spork/merge/spoon/SpoonMapping.java
@@ -5,6 +5,7 @@ import com.github.gumtreediff.matchers.MappingStore;
 import com.github.gumtreediff.tree.ITree;
 import com.github.gumtreediff.utils.Pair;
 import gumtree.spoon.builder.SpoonGumTreeBuilder;
+import se.kth.spork.util.GumTreeSpoonAstDiff;
 import spoon.reflect.declaration.CtElement;
 
 import java.util.ArrayList;
